-remote API -enemy & base generator fix

+++ Assets/Scripts/PlayerScore.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class PlayerScore : MonoBehaviour
 {
 
-    public static float playerScore = 0;
+    public static int playerScore =0;
     private Text scoreText;
     // Start is called before the first frame update
     void Start()
